% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstarcalc.R
\name{calc_pstar}
\alias{calc_pstar}
\title{Calculate P* from relative biomass and Council policy}
\usage{
calc_pstar(relB)
}
\arguments{
\item{relB}{Numeric. Single value of B/Bmsy (unitless). Must be a scalar, since it is used in `if` conditions.}
}
\value{
Numeric. P* value between 0 and 0.49.
}
\description{
Calculates P* from biomass relative to target (B/Bmsy) using the MAFMC P* policy. P* is 0 at or below B/Bmsy = 0.1 (fishery closed), rises linearly to 0.45 at B/Bmsy = 1, then to 0.49 at B/Bmsy = 1.5, where it plateaus.
}
\examples{
Pstar <- calc_pstar(1.55)
}
\author{
E Liljestrand, expanded code from M. Wilberg, originally 7-26-2011, updated with new MAFMC P* policy 5-10-2022
}
