% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareWHAMslides.R
\name{compareWHAMslides}
\alias{compareWHAMslides}
\title{Auto-generate model comparison slides}
\usage{
compareWHAMslides(
  plotPaths = NULL,
  format = "pptx",
  slideOut = NULL,
  slideTitle = NULL,
  fleetNames = NULL,
  fleetGroups = NULL,
  indexNames = NULL,
  indexGroups = NULL,
  modelNames = NULL,
  AIC = NULL,
  Gradient = NULL,
  Convergence = NULL,
  MohnsRho_F = NULL,
  MohnsRho_SSB = NULL,
  MohnsRho_R = NULL,
  Overfished = NULL,
  Overfishing = NULL,
  Rho_adjust = NULL
)
}
\arguments{
\item{plotPaths}{A vector of file paths pointing to plots folders created by plot_wham_output for all models to be compared, must end in "plots_png".}

\item{format}{A string indicating the type of format to use ("pptx" or "revealjs"), default = "pptx"}

\item{slideOut}{A file path for the resulting quarto document that formats the slides (must end in .qmd extension), no default.}

\item{slideTitle}{A string used for the slide title page}

\item{fleetNames}{A vector of fleet names in the order that they are provided in plotted results (assumes same fleets for all models, if this isn't the case use fleetGroup instead)}

\item{indexNames}{A vector of index names in the order that they are numbered in plotted results (assumes same indices for all models, if this isn't the case use indexGroup instead)}

\item{modelNames}{A vector of model names in the order that their plotted results are listed in plotPaths}

\item{AIC}{An optimal vector of AIC values for each model (order consistent with plotPaths) for inclusion in a summary table}

\item{Gradient}{An optimal vector of gradient convergence values for each model (order consistent with plotPaths) for inclusion in a summary table}

\item{Convergence}{An optimal vector of convergence values for each model (order consistent with plotPaths) for inclusion in a summary table}

\item{MohnsRho_F}{An optimal vector of Mohn's rho F values for each model (order consistent with plotPaths) for inclusion in a summary table}

\item{MohnsRho_SSB}{An optimal vector of Mohn's rho SSB values for each model (order consistent with plotPaths) for inclusion in a summary table}

\item{MohnsRho_R}{An optimal vector of Mohn's rho R values for each model (order consistent with plotPaths) for inclusion in a summary table}

\item{Overfished}{An optimal vector of overfished status for each model (order consistent with plotPaths) for inclusion in a summary table}

\item{Overfishing}{An optimal vector of overfishing status for each model (order consistent with plotPaths) for inclusion in a summary table}

\item{Rho_adjust}{An optimal vector of rho adjustment checks for each model (order consistent with plotPaths) for inclusion in a summary table

NOTE: Assumes stock and region names are consistent across multi-wham models OR compare single-wham model that lacks stock/region labels to a multi-wham model
WARNING: Code was updated to accommodate multi-wham results but the following settings have not been tested thoroughly following these edits: multiple stocks/regions/fleets, indices with non-standard names (i.e. not "index_1", "index_2"...), fleet and index groups (revert to version of function prior to July 2026 for functioning single-wham implementation)}

\item{fleetGroup}{A named list containing a vector of fleet numbers to compare across all models (e.g. if number of fleets or order differs across models specify which fleets should be compared), if provided these names are used to label slides instead of fleetNames. Length of all vectors in list must equal length of plotPaths, provide 0 if fleet does not exist for a given model}

\item{indexGroup}{A named list containing a vector of index numbers to compare across all models (e.g. if index order differs or an index is split in 2 for one model), if provided these names are used to label slides instead of indexNames. Length of all vectors in list must equal length of plotPaths, provide 0 if index does not exist for a given model}
}
\value{
A .qmd and output slides in the selected slideOut directory to compare the specified model runs
}
\description{
Compare diagnostics side by side for 2 WHAM models in pptx slides or 2+ in revealjs .html output.
}
\examples{
# Generalized file paths and labels are purely for demonstration and would need to be updated to run on individual machines
compareWHAMslides(
  plotPaths = c(
    here::here("model1_folder/plots_png"),
    here::here("model2_folder/plots_png")
  ),
  slideOut = here::here("compare_1_2_test.qmd"), slideTitle = "Testing title text",
  fleetNames = "US fleet", indexNames = c("DFO spring index", "NEFSC spring index ", "NEFSC fall index"), modelNames = c("model 1", "model 2")
)
# Example with models where default index numbering does not correspond
compareWHAMslides(
  plotPaths = c(
    here::here("model1_folder/plots_png"),
    here::here("model3_folder/plots_png")
  ),
  slideOut = here::here("compare_split_run.qmd"), slideTitle = "Testing title text",
  indexGroups = list(DFO_spring = c(1, 1), NEFSC_spring_Alb = c(2, 2), NEFSC_spring_Big = c(2, 3), NEFSC_fall_Alb = c(3, 4), NEFSC_fall_Big = c(3, 5), NEFSC_fall_BigOnly = c(0, 5))
)
}
