% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_stock_specs_dynamic.R
\name{export_stock_specs_dynamic}
\alias{export_stock_specs_dynamic}
\title{Export current stock specifications from StockEff to Excel}
\usage{
export_stock_specs_dynamic(
  dbconn = NULL,
  module,
  itis,
  stock,
  sa_year,
  assessment,
  schema = "STOCKEFF",
  sex = "NONE",
  out_dir = "./",
  max_rows = 1e+05,
  blacklist = c("I_CF_MONTHS_S", "I_OB_MONTHS_S", "I_CF_PROJECTED_SPECS_S",
    "I_OB_PROJECTED_SPECS_S", "I_SV_PROJECTED_SPECS_S"),
  whitelist = c("I_ASSESSMENT_NOTES_S"),
  STGonly = FALSE
)
}
\arguments{
\item{dbconn}{An active database connection object. If NULL, prompts the user for credentials.}

\item{module}{Character. One of "CF", "OB", or "SV".}

\item{itis}{Character. The species ITIS code.}

\item{stock}{Character. Stock abbreviation.}

\item{sa_year}{Numeric/Integer. Stock assessment year.}

\item{assessment}{Character. Assessment abbreviation.}

\item{schema}{Character. Target schema ("STOCKEFF" or "STOCKEFF_PRE_PROD"). Default is "STOCKEFF".}

\item{sex}{Character. Required for "SV" module parameters. Default is "NONE".}

\item{out_dir}{Character. Directory path to save the Excel file. Defaults to working directory.}

\item{max_rows}{Numeric. Safety threshold to prevent RAM crashes (default 100,000).}

\item{blacklist}{Character vector. Tables ending in '_S' to explicitly exclude from extraction.}

\item{whitelist}{Character vector. Tables ending in '_S' from other modules/global structures to explicitly include.}

\item{STGonly}{Logical. Set to TRUE to see results from staging form inputs for your stock (default FALSE)}
}
\value{
An xlsx file with current stock specifications.
}
\description{
Hardened Dynamic StockEff Specification Extractor - scrapes stockEff tables for current stock specifications
}
\examples{
\dontrun{
# For Commercial landings (Using existing connection)
export_stock_specs_dynamic(dbconn,
  module = "CF",
  itis = "172873",
  stock = "UNIT",
  sa_year = 2024,
  assessment = "MT2",
  out_dir = paste0(getwd(), "/ConfirmSpecs/")
)

# Dynamic login mode (If connection is NULL)
export_stock_specs_dynamic(
  dbconn = NULL,
  module = "CF",
  itis = "172873",
  stock = "UNIT",
  sa_year = 2024,
  assessment = "MT2",
  out_dir = paste0(getwd(), "/ConfirmSpecs/")
)
}
}
