% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge_multiWHAM.R
\name{fancy_scientific}
\alias{fancy_scientific}
\title{Helper function: fancy_scientific}
\usage{
fancy_scientific(l)
}
\description{
Helper function pulled from multi-wham compare_wham_models.R with some modifications to bridge from single- to multi-wham
}
