% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ismooth_functions.R
\name{fill_hole}
\alias{fill_hole}
\title{Fill a hole in a survey time series}
\usage{
fill_hole(data, year, season, howfill)
}
\arguments{
\item{data}{tibble with columns Year, Series, Index}

\item{year}{year to fill hole}

\item{season}{season to fill hole}

\item{howfill}{one of "avg", "pre", "post" determining how hole is filled}
}
\value{
tibble of same size with hole filled (note the order of the series can change)
}
\description{
fills a hole in the survey time series with average of pre and post values, or just the pre or post value
}
