% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge_multiWHAM.R
\name{get.ci}
\alias{get.ci}
\title{Helper function: get.ci}
\usage{
get.ci(
  par,
  se,
  p = 0.975,
  lo = 0,
  hi = 1,
  type = "I",
  k = 1,
  alpha.ci = 0.05,
  getci = TRUE,
  asap = FALSE
)
}
\description{
Helper function pulled from multi-wham compare_wham_models.R with some modifications to bridge from single- to multi-wham
}
