% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge_multiWHAM.R
\name{gg_facet_dims}
\alias{gg_facet_dims}
\title{Helper function: gg_facet_dims}
\usage{
gg_facet_dims(p)
}
\description{
Helper function pulled from multi-wham compare_wham_models.R with some modifications to bridge from single- to multi-wham
}
