% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invabccalc.R
\name{inv_ABC}
\alias{inv_ABC}
\title{Inverse function to calculate P* from ABC and OFL}
\usage{
inv_ABC(ABC, OFL, CV)
}
\arguments{
\item{ABC}{Numeric. Allowable biological catch, in the same units as `OFL`.}

\item{OFL}{Numeric. Overfishing limit for the stock, used as the lognormal median.}

\item{CV}{Numeric. Coefficient of variation of the OFL (unitless, e.g., 1.0).}
}
\value{
Numeric. P* value (probability between 0 and 1) implied by `ABC`.
}
\description{
Calculates the P* value implied by an ABC, given the OFL and the assumed CV of the OFL. This is the inverse of `ABC()` with respect to P*.
}
\examples{
pstar <- inv_ABC(11456, 13234, 1.0)
}
\author{
E Liljestrand, expanded code from M. Wilberg, originally 7-26-2011, updated with new MAFMC P* policy 5-10-2022
}
