% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_simtest_plots.R
\name{make_simtest_plots}
\alias{make_simtest_plots}
\title{Plots from a WHAM simulation self-test}
\usage{
make_simtest_plots(
  mod_selftest,
  saveall = FALSE,
  savepdf = FALSE,
  savepng = FALSE,
  myfilename_pdf = NULL,
  mypngprefix = NULL
)
}
\arguments{
\item{mod_selftest}{result of do_selftest function, contains annual true and estimated F, R, and SSB for a number of realizations}

\item{saveall}{Boolean (default=FALSE) when TRUE save all the possible plots, when FALSE save only single main plot}

\item{savepdf}{Boolean (default=FALSE) when TRUE save the selected plot(s) in a single pdf file}

\item{savepng}{Boolean (default=FALSE) when TRUE save the selected plot(s) as separate png files}

\item{myfilename_pdf}{(default=NULL) name (including filepath if so desired) for pdf to be created}

\item{mypngprefix}{(default=NULL) prefix added to default names for all saved png files (including filepath is so desired)}
}
\value{
list of 8 figures, main figure is first one
}
\description{
Automated creation of plots from a WHAM simulation self-test. Main plot shows mean, median, and 5%, 25%, 75%, 95% quantiles.
Optional additional plots show only mean with standard error, only median and quantiles, and numerous realizations.
Percent error (Est-True)/True computed for F, R, and SSB.
Requires ggplot2, dplyr.
}
