% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMTslides.R
\name{makeMTslides}
\alias{makeMTslides}
\title{Generate basic MT slides}
\usage{
makeMTslides(
  wham_model = NULL,
  outdir = NULL,
  slideName = "MTslides",
  stock_name = NULL,
  format = "pptx",
  fleet_names = NULL,
  index_names = NULL
)
}
\arguments{
\item{wham_model}{A final fitted wham model to use in generating MT slides, no default. NOTE: the loaded version of WHAM must match the wham_model (i.e. wham_model is a multi-wham model then multi-wham software must be loaded)}

\item{outdir}{A file path to store slides and wham plots in a folder titled "MTslides", no default.}

\item{slideName}{A string to use as the slide deck file name, default = "MTslides".}

\item{stock_name}{A string for the stock name, no default.}

\item{format}{A string indicating the type of format to use ("pptx" or "revealjs"), default = "pptx"}

\item{fleet_names}{A vector of fleet names in the order their data is provided to wham (also determines plot order), no default.}

\item{index_names}{A vector of index names in the order their data is provided to wham (also determines plot order), no default.}
}
\description{
Generate a standard set of MT slides given plots from a final WHAM model run
}
