% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge_multiWHAM.R
\name{plot.kobe.compare}
\alias{plot.kobe.compare}
\title{Helper function: plot.kobe.compare}
\usage{
\method{plot}{kobe.compare}(x, plot.opts)

\method{plot}{kobe.compare}(x, plot.opts)
}
\description{
Helper function pulled from multi-wham compare_wham_models.R with some modifications to bridge from single- to multi-wham

Helper function pulled from multi-wham compare_wham_models.R with some modifications to bridge from single- to multi-wham
}
