% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge_multiWHAM.R
\name{plot.M.compare}
\alias{plot.M.compare}
\title{Helper function: plot.M.compare}
\usage{
\method{plot}{M.compare}(x, compare.opts, plot.opts)
}
\description{
Helper function pulled from multi-wham compare_wham_models.R with some modifications to bridge from single- to multi-wham
}
