% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge_multiWHAM.R
\name{plot.rel.compare}
\alias{plot.rel.compare}
\title{Helper function: plot.rel.compare}
\usage{
\method{plot}{rel.compare}(x, plot.opts)
}
\description{
Helper function pulled from multi-wham compare_wham_models.R with some modifications to bridge from single- to multi-wham
}
