% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge_multiWHAM.R
\name{plot.selectivity.compare}
\alias{plot.selectivity.compare}
\title{Helper function: plot.selectivity.compare}
\usage{
\method{plot}{selectivity.compare}(x, plot.opts, type = "fleet")
}
\description{
Helper function pulled from multi-wham compare_wham_models.R with some modifications to bridge from single- to multi-wham
}
