% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge_multiWHAM.R
\name{plot.SSB.F.R.compare}
\alias{plot.SSB.F.R.compare}
\title{Helper function: plot.SSB.F.R.compare}
\usage{
\method{plot}{SSB.F.R.compare}(x, compare.opts, plot.opts)
}
\description{
Helper function pulled from multi-wham compare_wham_models.R with some modifications to bridge from single- to multi-wham
}
