% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge_multiWHAM.R
\name{plot.tile.compare}
\alias{plot.tile.compare}
\title{Helper function: plot.tile.compare}
\usage{
\method{plot}{tile.compare}(x, compare.opts, plot.opts, type = "selAA")
}
\description{
Helper function pulled from multi-wham compare_wham_models.R with some modifications to bridge from single- to multi-wham
}
