% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge_multiWHAM.R
\name{plot.timeseries.compare}
\alias{plot.timeseries.compare}
\title{Helper function: plot.timeseries.compare}
\usage{
\method{plot}{timeseries.compare}(df, x, plot.opts)
}
\description{
Helper function pulled from multi-wham compare_wham_models.R with some modifications to bridge from single- to multi-wham
}
