% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ismooth_functions.R
\name{wrapIsmooth}
\alias{wrapIsmooth}
\title{Standardize, plot, and run Ismooth}
\usage{
wrapIsmooth(
  data,
  YearStart,
  YearEnd,
  myseries,
  mytitle,
  mynarm = FALSE,
  print_plots = FALSE
)
}
\arguments{
\item{data}{tibble with columns Year, Series, Index}

\item{YearStart}{data filtered to years >= this value}

\item{YearEnd}{data filtered to years <= this value}

\item{myseries}{vector of series names in quotes that subsets the Series column}

\item{mytitle}{character string that will appear on top of plot_standardized_and_avg plot and Ismooth plot}

\item{mynarm}{Boolean to remove NA values when calculating annual average index, default=FALSE}

\item{print_plots}{when TRUE plots standardized and avg plot as well as Ismooth time series plot, default=FALSE}
}
\value{
list with std_data tibble, avg tibble, standardized and average plot, and Ismooth results
}
\description{
wrapper function to standardize and combine time series, optionally plot, and run Ismooth
}
