% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abccalc.R
\name{ABC}
\alias{ABC}
\title{Function to calculate the ABC using a lognormal distribution}
\usage{
ABC(OFL, relB, CV)
}
\arguments{
\item{OFL}{OFL for the stock}

\item{relB}{B/Bmsy}

\item{CV}{Coefficient of variation for OFL}
}
\value{
ABC - allowable biological catch
}
\description{
Calculates ABC using inputs of the OFL, biomass relative to target and assumed CV for OFL
}
\examples{
catch <- ABC(12345,1.11,0.6)
}
\author{
E Liljestrand, expanded code from M. Wilberg, originally 7-26-2011, updated with new MAFMC P* policy 5-10-2022
}
