% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asm.ci.R
\name{asm.ci}
\alias{asm.ci}
\title{asm.ci}
\usage{
asm.ci(x, cv.x, bounds = 95)
}
\arguments{
\item{x}{A lognormal variable (or vector of variables), no default.}

\item{cv.x}{The cv of variable x (or vector of CVs if multiple variables provided), no default.}

\item{bounds}{The confidence bound to calculate, default = 95.}
}
\value{
A data.frame containing lower (lci) and upper (uci) confidence bounds
}
\description{
Calculate approximate 95% confidence intervals based on log normal variable and its CV. Used in Dan's autoreporting as default for survey index CI calculations.
}
