% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borrow_ages.R
\name{borrow_ages}
\alias{borrow_ages}
\title{borrow_ages}
\usage{
borrow_ages(
  missing_ages,
  CF_ages,
  SV_ages,
  OB_ages,
  borrow_procedure = "seqCFSV_S_A",
  borrow_2020_option = FALSE
)
}
\arguments{
\item{missing_ages}{A table containing length-age pairs that lack age information to fill via borrowing_procedure !!! probably want to describe columns/content better}

\item{CF_ages}{A table of commercial ages to use in borrowing !!! describe what stockEff products used to generate these and other ages table arguments}

\item{SV_ages}{A table of survey ages to use in borrowing}

\item{OB_ages}{PLACEHOLDER FOR NOW: A table of observer ages to use in borrowing}

\item{borrow_procedure}{A string specifying the borrowing logic to apply to fill age gaps for ALK, options include:
\itemize{
  \item{"seqCFSV_S_A" - Default, Borrow: 1) CF within semester, 2) CF within year, 3) SV within semester, 4) SV within year} 
  \item{"seqSA_CF_SV" - Borrow: 1) CF within semester (i.e. pool mkt categories), 2) SV within semester, 3) CF within year, 4) SV within year} 
}}

\item{borrow_2020_option}{A string specifying how 2020 borrowing is handled, options include: 
\itemize{
  \item{Default uses borrow_procedure to fill 2020 gaps, in most cases this results in no gap filling because borrowing does not occur across years}
  \item{"none" = do not borrow to fill 2020 age gaps}
  \item{"combCFSV_1921" = Combine CF and SV data, fill 2020 using all observations from 2019 and 2021}
}}
}
\value{
A list containing:
\itemize{
  \item{all_fills - A table containing borrowed data used to fill gaps for missing_ages}
  \item{remaining_gaps - A table of missing_ages for which gaps could not be filled using borrow_procedure logic}
}
}
\description{
borrow_ages
}
