% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge_multiWHAM.R
\name{bridge_multiWHAM}
\alias{bridge_multiWHAM}
\title{Bridge from single to multi-wham}
\usage{
bridge_multiWHAM(
  mod_single = NULL,
  mod_multi = NULL,
  fdir = here::here(),
  table.opts = NULL,
  plot.opts = NULL
)
}
\arguments{
\item{mod_single}{A model object fitted in single/regular WHAM, no default.}

\item{mod_multi}{A model object fitting the same model specification in multi-WHAM, no default.}

\item{fdir}{A file path where comparison table and plots will be stored, default is project directory.}

\item{table.opts}{Table options that roughly match compare_wham_models() options, no guarantee that all options operate, default = NULL.}

\item{plot.opts}{Plot options that roughly match compare_wham_models() options, no guarantee that all options operate, default = NULL.}
}
\value{
A list containing the following:
\itemize{
  \item{sdrep_summary - Summary of joined sdrep objects, repeated labels indicate differences in estimated values (rounded to 3 decimals)}
  \item{different_pars - Vector of parameter names that had different estimates between the single and multi-wham models}
  \item{tab - Table of AIC and Mohn's rho values by model}
  \item{out - Comparison plots}
  \item{sdrep_diff - Table of parameter estimates that differ between single and multi-wham versions of the model}
  #!!! Need to add a check here for the bias correction setting since the defaults differ and this is an easy change to forget about when switching
}
}
\description{
Compares a single-wham model to a multi-wham model with 1 stock and 1 area to bridge from single-wham to multi-wham (allows 1 stock/region models to take advantage of additional features in multi-wham software (e.g. initial stock conditions))
}
\examples{
mod_single <- readRDS(file = here::here("Bridge_runs", "9_bridge_reviseCV_NEFSC", "9_bridge_model.rds"))
mod_multi <- readRDS(file = here::here("Bridge_runs", "12_bridge_updateWHAM", "12_bridge_model.rds"))
bridge_multiWHAM(mod_single = mod_single, mod_multi = mod_multi, fdir = here::here("Bridge_runs/12_bridge_updateWHAM"), table.opts= list(save.csv = TRUE))

}
