% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.rho.adj.ests.R
\name{calc.rho.adj.ests}
\alias{calc.rho.adj.ests}
\title{calc.rho.adj.ests}
\usage{
calc.rho.adj.ests(series = NULL, rho = NULL)
}
\arguments{
\item{series}{A data frame containing the following columns:
\itemize{
\item est -  Estimated time series
\item lo_90 - Lower 90\% CI for estimates
\item hi_90 - Upper 90\% CI for estimates
}}

\item{rho}{Mohn's rho value associated with the time series.}
}
\value{
A data table containing original and rho-adjusted time series with accompanying 90\% CI.
}
\description{
Calculate rho-adjusted time series based on 90\% CI
}
