% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.uncertainty.R
\name{calc.uncertainty}
\alias{calc.uncertainty}
\title{Calculate uncertainty}
\usage{
calc.uncertainty(log.est = NULL, log.se = NULL)
}
\arguments{
\item{log.est}{An estimated time series in log space, no default.}

\item{log.se}{A vector of log standard errors corresponding to the estimated time series, no default.}
}
\value{
A data frame containing:
\itemize{
\item{est - Time series of estimated values}
\item{se - Accompanying standard errors}
\item{CV - Accompanying CVs}
\item{lo_95 - Lower 95\% confidence intervals}
\item{hi_95 - Upper 95\% confidence intervals}
\item{lo_90 - Lower 90\% confidence intervals}
\item{hi_90 - Upper 90\% confidence intervals}
}
}
\description{
Calculate CV, 90\% and 95\% CI given a model estimate and standard error.
}
