% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_caa_waa.R
\name{calc_caa_waa}
\alias{calc_caa_waa}
\title{calc_caa_waa}
\usage{
calc_caa_waa(
  ALK = NULL,
  stockEff_mode = "prod",
  stockEff_module = NULL,
  species_itis = NULL,
  connection = NULL
)
}
\arguments{
\item{ALK}{A long-format table containing the following columns:
\itemize{
  \item{YEAR}
  \item{LENGTH - Length bin}
  \item{AGE - Age bin}
  \item{PROP - proportion-at-age for a given length and YEAR}
  \item{BLOCK_ID - Block ID used to match ALK to expanded lengths from stockEff}
}}

\item{stockEff_mode}{String specifying version of stockEff to query, options include: "test", "prod". Default = "prod"}

\item{stockEff_module}{A string indicating the module for which products will be pulled, no default. Options include:
\itemize{
  \item{"survey" - Correspond to SV tab in stockEff}
  \item{"commercial" - Correspond to CF tab in stockEff}
  \item{"observer" - Correspond to OB tab in stockEff}
  \item{Nothing yet available for MRIP tab (as of 3/6/24)}
}}

\item{species_itis}{Species itis code}

\item{connection}{A DBI or ROracle connection object to the NEFSC Oracle database.
If NULL (default), the function will prompt for credentials via rstudioapi
and manage the connection/disconnection automatically. If a connection
is provided, the user is responsible for closing it after the function executes.}
}
\description{
Calculate annual catch- and weight-at-age, using expanded lengths from stockEff and ALK with time blocking that matches stockEff blocking

NOTE: If you receive the following error try setting connection = NULL and rerun function to re-establish oracle connection:
      Error: unable to find an inherited method for function ‘dbGetQuery’ for signature ‘conn = "NULL", statement = "character"’
}
