% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstarcalc.R
\name{calc_pstar}
\alias{calc_pstar}
\title{Function to calculate P* FROM RELATIVE BIOMASS AND COUNCIL POLICY}
\usage{
calc_pstar(relB)
}
\arguments{
\item{relB}{B/Bmsy}
}
\value{
P*
}
\description{
Calculates P* using input of the Biomass relative to target
}
\examples{
Pstar <- calc_pstar(1.55)
}
\author{
E Liljestrand, expanded code from M. Wilberg, originally 7-26-2011, updated with new MAFMC P* policy 5-10-2022
}
