% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ismooth_functions.R
\name{clean_stockeff_survey_data}
\alias{clean_stockeff_survey_data}
\title{clean_stockeff_survey_data}
\usage{
clean_stockeff_survey_data(
  mydata,
  mypurpose,
  myseasons,
  myindextype,
  myfills = NULL
)
}
\arguments{
\item{mypurpose}{code for survey type in StockEff (NEFSC bottom trawl survey = 10)}

\item{myseasons}{vector of seasons, e.g., c("SPRING", "FALL")}

\item{myindextype}{character string "Biomass (kg/tow)" or "Abundance (numbers/tow)"}

\item{myfills}{tibble of Year, Series, Howfill (latter can be "avg", "pre", "post"), set to NULL if none, default=NULL}

\item{data}{tibble with columns YEAR, SEASON, PURPOSE_CODE, INDEX_TYPE, INDEX (can be read from StockEff csv file, other columns allowed)}
}
\value{
tibble of Year, Series, Index with Series having up to FILLED_SPRING, FALL, FILLED_FALL, FILLED_FALL_LAGGED levels
}
\description{
cleans StockEff survey data by filtering on seasons and index type as well as filling holes
}
