% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.brp.text.R
\name{create.brp.text}
\alias{create.brp.text}
\title{Function to create the brp text}
\usage{
create.brp.text(brp.name = NULL, round.digits = 3, brp.table = NULL)
}
\arguments{
\item{brp.name}{A string for the type of BRP, options include: 'SSBproxy', 'MSYproxy', and 'Fproxy', no default.}

\item{round.digits}{The number of digits to round BRPs, default = 3.}

\item{brp.table}{A table containing the following columns:
\itemize{
  \item{BRP - Name of the BRP}
  \item{est - BRP estimates}
  \item{lo_95 - Lower 95\% confidence bound}
  \item{hi_95 - Upper 95\% confidence bound}
  \item{source - Model source, either "prior" for prior assessment or "MT" for current assessment}
}}
}
\value{
A string for the BRP text including 95% CI.
}
\description{
Function to create the brp text
}
