% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_selftest.R
\name{do_selftest}
\alias{do_selftest}
\title{do_selftest}
\usage{
do_selftest(mod, n_sim)
}
\arguments{
\item{mod}{wham model created with `fit_wham()`}

\item{n_sim}{number of simulations to perform}
}
\value{
data.frame of annual true and estimated F, R, and SSB
}
\description{
Perform simulation self-test of a WHAM model.
}
