% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_term_index_retro.R
\name{drop_term_index_retro}
\alias{drop_term_index_retro}
\title{missing terminal year indices}
\usage{
drop_term_index_retro(model = NULL, drop = NULL)
}
\arguments{
\item{model}{A WHAM rds output file from a run that included NAA random effects and a retrospective. The rds cannot include projections.}

\item{drop}{A list with two elements (indices and index_paa) that gives the numeric value indicating which indices and age comps to drop from the terminal years}
}
\value{
A list containing the results of each fit. Also saves a png plot of NAA devs from each fit and the SSB retro where the selected index data are missing from each terminal year. Saved to the working directory
}
\description{
drops some or all of the indices from the terminal year of each peel of retrospective to see how the NAA devs and retro change. Reliant on the fit_hindcast and plotNAAfxn.
}
\examples{
#drops all indices from a fit
droptermIndex=drop_term_index_retro(model=mod,drop=list(indices=1:mod$input$data$n_indices,index_paa=1:mod$input$data$n_indices)) 
drop whatever index 2 is from a fit
droptermIndex=drop_term_index_retro(model=mod,drop=list(indices=2,index_paa=2)) 
}
