% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillCF.R
\name{fillCF}
\alias{fillCF}
\title{Fill CF specification template for STOCKEFF}
\usage{
fillCF(
  species_itis = NULL,
  stock_abbrev = NULL,
  assessment_abbrev = NULL,
  assessment_year = NULL,
  data_years = NULL,
  sex_type = "NONE",
  tab_blocks = list(autofillLW = FALSE, mkt_missing = "keep", check_missing = FALSE,
    block_opt = "ANNUAL", minLenSample = 1),
  tab_prorate = NULL,
  tab_regions = NULL,
  tab_lw = NULL,
  tab_sex = NULL,
  tab_impute = NULL,
  tab_alk = list(fill_alk = FALSE, borrow_alk_years = FALSE, source = "both"),
  tab_exclusions = list(len_range = NULL, age_range = NULL),
  schema = "STOCKEFF",
  outfile = NULL,
  conn = NULL
)
}
\arguments{
\item{species_itis}{6 digit ITIS identifying the species for which data should be pulled, no default.}

\item{stock_abbrev}{A string describing the stock abbreviation assigned by stock efficiency (this shows up in the web address for STOCKEFF products), no default.}

\item{assessment_abbrev}{A string describing the assessment abbreviation, no default. Common options include "MT", "RT", "OA"}

\item{assessment_year}{The year in which the assessment will be reviewed, no default.}

\item{data_years}{The range of years for which catch data should be formatted for STOCKEFF CF excel template, no default.}

\item{sex_type}{A string describing the sex type for which data should be pulled (this shows up in the web address for STOCKEFF products), default = "NONE". Common options include "NONE", "MALE", "FEMALE", "UNSEXED"}

\item{tab_blocks}{A list providing the following information to populate the BLOCKS tab: (NOTE: LW_ID is populated by identifying the LW whose associated start and end months completely overlap with the block, otherwise an NA is returned)
\itemize{
  \item{nespp4 - A named vector of market categories to include in block tab, where names correspond to the nespp4 codes. No default. Market categories included in the stock definition but excluded here are by default prorated (see tab_prorate for specifics)}
  \item{dropMarketYr - A table containing the nespp4 market category ("NESPP4" column) to "roll up" (i.e. drop) in the corrsponding year column("YEAR"), if provided landings are redistributed to other categories for that year/region (see "roll up" in Adding and Updating Stocks in StockEff documentation)}
  \item{block_opt - A string for the following options to specify blocks for each market category and year: 
          if = NULL (default), one block for every L-W relationship in tab_lw, 
          if = "MONTH", one block for each month (useful for checking minimum number of lengths exceeded for larger blocks), 
          if = "QUARTER", one block for each quarter,
          if = "SEMESTER", one block for each semester,
          if = "ANNUAL", one block per year for each market category
          if = "checkLENGTHS", Assumes monthly blocks were submitted to STOCKEFF, checks the number of monthly length samples from biosamp_summary.csv to determine smallest block supported by data for each market category and year. Must also provide minLenSample argument if this option is used.}
  \item{minLenSample - A number specifying the minimum number of length sampels required to support a given block for a year and market category, used in length imputation check even if block_opt != checkLENGTHS, default = 1}
  \item{autofillLW - A boolean, if TRUE fills LW_ID based on complete overlap with LW months and years, if FALSE leaves NAs for blocks without a perfect match to LW month and year definition, default = FALSE}
  \item{mkt_missing - A string to govern years where 1+ market categories not length sampled, default = "keep":
           if = "drop" then these years dropped and NAA is not estimated
           if = "keep" then return NAs for BLOCK_TYPE in these years}
  \item{check_missing - A boolean, if TRUE check for missing market categories in data_years, also check if is.na(mkt_missing) == FALSE (i.e. if years with missing market categories will be dropped or further processed)}
}}

\item{tab_prorate}{A list providing the following information to populate the PRORATE tab:
\itemize{
  \item{turnOff - A vector of years for which unclassified landings should NOT be prorated, default prorates for all data_years.}
}}

\item{tab_regions}{A list providing the following information to populate the REGIONS tab:
\itemize{
  \item{stat_areas - A list of stat area vectors by region, stat areas should be provided as strings. Example: list("1" = c("511", "512"))}
  \item{region_custom - An optional data.frame/matrix for complex cases. 
        Example: data.frame(AREA = c(511, 512, 511), 
                           REGION_ID = c("1", "1", "2"), 
                           YEAR = c(2020, 2020, 2021))
        This is useful if an AREA changes REGION_ID over time.}
}}

\item{tab_lw}{A matrix or data.frame containing the following named columns of information for each LW data source. Currently only supports blocks with types, start and end months that match LW_PARAMS and are consistent across all years
\itemize{
  \item{ALPHA - Alpha parameter for each LW relationship}
  \item{BETA - Beta parameter for each LW relationship}
  \item{SOURCE - Description of LW parameter and link to source document}
  \item{LW_TYPE - Type of LW relationship, options can include "SEMESTER", "ANNUAL", "QUARTER"}
  \item{LW_ID - ID for each row of the table}
  \item{MONTH_START - First month that LW_ID row may be used, processed in tab_BLOCKS and ultimately dropped from tab_LW_PARAMS}
  \item{MONTH_END - Last month that LW_ID row may be used, processed in tab_BLOCKS and ultimately dropped from tab_LW_PARAMS}
  \item{YEAR_START - First year that LW_ID row may be used, processed in tab_BLOCKS and ultimately dropped from tab_LW_PARAMS}
  \item{YEAR_END - Last year that LW_ID row may be used, processed in tab_BLOCKS and ultimately dropped from tab_LW_PARAMS}
}}

\item{tab_sex}{A list providing the following information to populate the SEX_TYPE tab:
\itemize{
  \item{start_yr - The start year for CF data, default = first year in data_years.}
  \item{end_yr - The end year for CF data, default = last year in data_years.}
}}

\item{tab_impute}{A list providing rules for filling sampling gaps in the LENGTH_IMPUTATIONS tab:
\itemize{
  \item{borrow_from_region - A named list defining spatial donor relationships. 
        Example: list("5" = "6") means if Region 5 has a sampling hole, 
        borrow the length-frequency distribution from Region 6 for that same 
        Year and Market Category.}
  \item{borrow_from_time - A boolean. If TRUE, triggers the constrained temporal 
        hierarchy if spatial borrowing fails or is not defined. 
        The hierarchy follows: Same Semester (for Quarter holes) > 
        Same Year (for Semester holes) > Adjacent Years (for Year holes).}
}}

\item{tab_alk}{A list providing rules for filling gaps in the ALK_HOLES tab:
\itemize{
  \item{fill_alk - A boolean. If TRUE, identifies lengths present in landings that 
        lack corresponding age data and attempts to fill them using hierarchical borrowing. Default = FALSE.}
  \item{source - A vector of strings identifying data sources for age data gap filling,options are: "commercial", "survey". Default = "commercial", MUST contain "commercial" or ALK borrowing will trigger error} 
  \item{borrow_procedure - A string specifying the borrowing logic to apply to fill age gaps for ALK:
       
  }
  \item{borrow_2020_option - A string specifying how 2020 borrowing is handled: 
           "none" = do not borrow to fill 2020 age gaps (default)
           "match_borrow_procedure" = use same logic as borrow_procedure argument,
           
       } 
  \item{survey_region - Optional dataframe containing STRATUM and REGION_ID columns specifying the region 
       assignment for each survey stratum. Only required if multiple regions AND survey data used for 
       borrowing (i.e. "survey" in source list). Default assigns all survey strata to region 1}   
}}

\item{tab_exclusions}{An optional (default = NULL) list defining automated filters to remove non-representative biological records:
\itemize{
  \item{len_range - A numeric vector of length 2, e.g., c(10, 100). Any fish with a LENGTH 
        strictly less than the first value or strictly greater than the second value 
        will be added to the EXCLUSIONS tab.}
  \item{age_range - A numeric vector of length 2, e.g., c(1, 20). Any fish with an AGE 
        outside this range will be excluded. This is a powerful tool for removing 
        extreme outliers that may be age-reading or data-entry errors.}
}}

\item{schema}{Set to "STOCKEFF_PRE_PROD" if using pre production data for assembly. Default is "STOCKEFF"}

\item{outfile}{A string for the final exported file name, does NOT need to include .xlsx extention. No default.}

\item{conn}{A DBI or ROracle connection object to the NEFSC Oracle database. 
       If NULL (default), the function will prompt for credentials via rstudioapi 
       and manage the connection/disconnection automatically. If a connection 
       is provided, the user is responsible for closing it after the function executes.
        
\itemize{
\item{NOTE: tab_blocks$block_opt = checkLENGTHS will NOT work unless connected to the VPN}
\item{NOTE: ALK_HOLES, and EXCLUSIONS tabs are optionally populated by this function and may need to be filled externally if tab_alk and tab_exclusions arguments are not used to fill gaps.}
\item{NOTE: If you receive the following error, then the .xlsx file by the provided name already exists and can't be overwritten: [ERROR] workbook_close(): Error creating 'outfile.xlsx'. System error = Permission denied Error: Error in libxlsxwriter: 'Error creating output xlsx file. Usually a permissions error.'}
\item{NOTE: If you find NAs in the REGION_ID for the BLOCKS tab, then there was a stat area in the {schema}.mv_cf_stock_data_length_o product that was not provided in the tab_regions$stat_areas argument. Will also result in REGION_ID = NA in ALKHOLES tab.}
}}
}
\value{
A populated .xlsx template and a list containing the following:
\itemize{
  \item{mkt_missing_yr - A table of market categories years for which 1+ market categories were not sampled where 1+ market categories were not sampled}
  \item{regions_missing - Table of AREAs that appear in oracle data but not in tab_regions$stat_area argument for this function (may be missing blocking)}
}
}
\description{
Fill CF specification template for STOCKEFF
}
\examples{
\dontrun{
# Example call for GB Cod using spatial and temporal borrowing
fillCF(
  species_itis = 164712,
  stock_abbrev = "GBK",
  assessment_abbrev = "MT",
  assessment_year = 2024,
  data_years = 2019:2023,
  tab_regions <- list(
    stat_areas = list(
      "1" = c(464, 522, 523, 524, 525, 561, 562, 551, 552, 542, 543)
    ))
# for 2 regions
# tab_regions <- list(
#     stat_areas = list(
#       "1" = c(511, 512, 513),
#       "2" = c(521, 522, 525)
#       )
#     )
 tab_blocks <-  list(
  nespp4 = c(LARGE = 811, Market = 813, Scrod = 814),
  minLenSample = 100,
  block_opt = "checkLENGTHS",
  autofillLW = TRUE,
  mkt_missing = "keep",
  autofillLW = T
 ) 
  tab_impute = list(
    borrow_from_region = list("1" = "2"),
    borrow_from_time = TRUE
  ),
  outfile = "GB_Cod_Specifications"
)


#mackerel example using exclusions
species_itis = 172414
stock_abbrev = "UNIT"
assessment_abbrev = "MT"
assessment_year = 2025
data_years = 2018:2019 # small subset for testing
sex_type = "NONE"
tab_regions <- list(
  stat_areas = list(
    "1" = c(464,465,467,468,510,511,512,513,514,515,520,521,522,523,524,525,526,
            530,533,534,537,538,539,541,542,543,551,552,560,561,562,600,610,611,
            612,613,614,615,616,621,622,623,624,625,626,627,628,629,631,632,633,
            634,635,636,637,638,639,640,0
    )
  )
)

#extract lw eq. from STOCKEFF
lwQry <- glue::glue("select * from stockeff_pre_prod.i_cf_stock_lw_params_s where species_itis = {species_itis}
                    and stock_abbrev = '{stock_abbrev}' and assessment_abbrev = '{assessment_abbrev}'") 
fillLW <- ROracle::dbGetQuery(conn,lwQry)
tab_lw = data.frame(ALPHA = fillLW$ALPHA,
                    BETA = fillLW$BETA,
                    SOURCE = fillLW$SOURCE,
                    LW_TYPE = fillLW$LW_TYPE,
                    LW_ID = fillLW$LW_ID,
                    MONTH_START = c(1,7),
                    MONTH_END = c(6,12),
                    YEAR_START = rep(2000,2),
                    YEAR_END = rep(2024,2))
tab_blocks <-  list(nespp4 = c(MEDIUM = 2123, SMALL = 2122, LARGE = 2121, UNCLASSIFIED =2120),
                    minLenSample = 500,
                    block_opt = "checkLENGTHS",
                    autofillLW = TRUE,
                    mkt_missing = "keep")

tab_impute = list(
  #borrow_from_region = list("1" = "2", "2" = "1"), #only valid for stocks with more than one region!
  borrow_from_time = TRUE # Flag to trigger the temporal fallback
)

tab_exclusions = list(len_range = c(10,35))

outfile = file.path("ExampleCFspecs")
if(!dir.exists("ExampleAutoCFspecs")) dir.create("ExampleAutoCFspecs")
outfile = file.path("ExampleAutoCFspecs",glue::glue("CF_specification_load_{species_itis}_{stock_abbrev}_{assessment_year}"))

tab_prorate = NULL
tab_sex = NULL

fillCF(species_itis = species_itis,
       stock_abbrev = stock_abbrev,
       assessment_abbrev = assessment_abbrev,
       assessment_year = assessment_year,
       data_years = data_years,
       sex_type = sex_type,
       tab_regions = tab_regions, 
       tab_lw = tab_lw,           
       tab_blocks = tab_blocks,
       outfile = outfile,
       tab_prorate = tab_prorate,
       tab_sex = tab_sex,
       tab_impute = tab_impute,
       tab_alk = tab_alk,
       tab_exclusions = tab_exclusions,
       conn = conn
)

}


}
