% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillOB.R
\name{fillOB}
\alias{fillOB}
\title{Fill OB specification template for stockEff}
\usage{
fillOB(
  species_itis = NULL,
  stock_abbrev = NULL,
  assessment_abbrev = NULL,
  assessment_year = NULL,
  data_years = NULL,
  sex_type = "NONE",
  tab_blocks = list(autofillLW = FALSE, check_missing = FALSE),
  tab_prorate = NULL,
  tab_regions = NULL,
  tab_lw = NULL,
  tab_sex = NULL,
  tab_nespp4_flt = list(pre_CAMS = NULL, post_CAMS = NULL),
  tab_svage_def = NULL,
  tab_age_data_source = NULL,
  outfile = NULL
)
}
\arguments{
\item{species_itis}{6 digit ITIS identifying the species for which data should be pulled, no default.}

\item{stock_abbrev}{A string describing the stock abbreviation assigned by stock efficiency (this shows up in the web address for STOCKEFF products), no default.}

\item{assessment_abbrev}{A string describing the assessment abbreviation, no default. Common options include "MT", "RT", "OA"}

\item{assessment_year}{The year in which the assessment will be reviewed, no default.}

\item{data_years}{The range of years for which discard data should be formatted for stockEff OB excel template, no default.}

\item{sex_type}{A string describing the sex type for which data should be pulled (this shows up in the web address for STOCKEFF products), default = "NONE". Common options include "NONE", "MALE", "FEMALE", "UNSEXED"}

\item{tab_blocks}{A list providing the following information to populate the BLOCKS tab: (NOTE: LW_ID is populated by identifying the LW whose associated start and end months completely overlap with the block, otherwise an NA is returned)
\itemize{
  \item{block_opt - A string for the following options to specify blocks for each market category and year: 
          if = NULL (default), one block for every L-W relationship in tab_lw, 
          if = "MONTH", one block for each month (useful for checking minimum number of lengths exceeded for larger blocks), 
          if = "QUARTER", one block for each quarter,
          if = "SEMESTER", one block for each semester,
          if = "ANNUAL", one block per year for each market category
          if = "checkLENGTHS", Assumes monthly blocks were submitted to stockEff, checks the number of monthly length samples from biosamp_summary.csv to determine smallest block supported by data for each market category and year. Must also provide minLenSample argument if this option is used.}
  \item{minLenSample - A number specifying the minimum number of length samples required to support a given block for a year and gear, only required if tab_blocks$block_opt == "checkLENGTHS", if used then code checks for all data_years}
  \item{autofillLW - A boolean, if TRUE fills LW_ID based on complete overlap with LW months and years, if FALSE leaves NAs for blocks without a perfect match to LW month and year definition, default = FALSE}
  \item{check_missing - A boolean, if TRUE check for years in data_years with missing regions}
}}

\item{tab_prorate}{A list providing the following information to populate the PRORATE tab:
\itemize{
  \item{turnOff - A vector of years for which unclassified landings should NOT be prorated, default prorates for all data_years.}
}}

\item{tab_regions}{A list providing the following information to populate the REGIONS tab: (NOTE: code is less well tested with >1 region)
\itemize{
  \item{stat_areas - A list of stat area vectors by region, no default. Example: list(c(510, 511), c(512, 513)) assumes 2 regions each containing 2 stat areas}
  \item{area_fished - A list of area fished by stat area region, no default. Example: list(c("GOMGB"), c("SNE)) assumes 2 areas fished for the regions defined in stat_areas example}
  \item{region_id !!! tells you what region id(s) assigned to each area_fished}
  \item{gear_area_fished !!! tells you what NESPP4_FLT gear codes are associated with each region_id for each area fished}
  \item{survey_strata - A list of survey strata by region, no default. Used to connect area_fished to survey strata for SVAGE_DEFINITION tab Example: list(c(01130, 01140), c(01380, 01390, 01400))} assumes 2 regions with different survey strata assigned to each
  \item{region_custom - A optional custom matrix containing "REGION_ID", "AREA", and "YEAR" columns, not required if stat_areas object provided !!! does not currently work but could be built out in future}
}}

\item{tab_lw}{A matrix or data.frame containing the following named columns of information for each LW data source. Currently only supports blocks with types, start and end months that match LW_PARAMS and are consistent across all years
\itemize{
  \item{ALPHA - Alpha parameter for each LW relationship}
  \item{BETA - Beta parameter for each LW relationship}
  \item{SOURCE - Description of LW parameter and link to source document}
  \item{LW_TYPE - Type of LW relationship, options can include "SEMESTER", "ANNUAL", "CUSTOM"}
  \item{LW_ID - ID for each row of the table}
  \item{MONTH_START - First month that LW_ID row may be used, processed in tab_BLOCKS and ultimately dropped from tab_LW_PARAMS}
  \item{MONTH_END - Last month that LW_ID row may be used, processed in tab_BLOCKS and ultimately dropped from tab_LW_PARAMS}
  \item{YEAR_START - First year that LW_ID row may be used, processed in tab_BLOCKS and ultimately dropped from tab_LW_PARAMS}
  \item{YEAR_END - Last year that LW_ID row may be used, processed in tab_BLOCKS and ultimately dropped from tab_LW_PARAMS}
  \item{BLOCK_ID - Block ID associated with each LW equation, ultimately dropped from tab_LW_PARAMS (e.g. seasonal SPRING LW = block 1, FALL LW = 2, but an annual LW with only one time block = 1)}
}}

\item{tab_sex}{A list providing the following information to populate the SEX_TYPE tab:
\itemize{
  \item{start_yr - The start year for OB data, default = first year in data_years.}
  \item{end_yr - The end year for OB data, default = last year in data_years.}
}}

\item{tab_nespp4_flt}{A named list of "pre_CAMS" and "post_CAMS" settings providing lists of the following information to populate the NESPP4_FLT tab for each pre/post CAMS area:
\itemize{
  \item{nespp4_FLT - A vector of market categories corresponding to fleets (as done in BIOSTAT) to include in block and region tabs. No default.}
  \item{gear_group - A list of gear groups (stockEff or CAMS) included in each fleet definition (if multiple gear group codes included in a single fleet definition then these should be grouped into a vector and will have a separate line in NESPP4_FLT tab tieing it into the group). E.g. list(058, c(050, 0_trawlbot, 057)) would reflect 2 fleets with 3 gear groups combined in the second fleet}
  \item{NEGEAR_group - A list of NEGEAR codes that are included in each of the nespp4_FLT defintions (only required if tab_blocks$block_opt == "checkLENGTHS" to query OBDBS lengths). E.g. for the above gear_group example, the corresponding NEGEAR_group argument should be: list(058, c(050, 353, 057, 051, 054, 055, 056, 059, 150, 350, 351)) based on CAMS gear groups}
  \item{meshsize_abbrev - A vector or mesh size abbreviations corresponding to the gear group (if not a trawl gear set to "ALL"), common abbreviation are "LG", "SM", "ALL" (i.e. each unique combination of gear_group and meshsize_abbrev represents a unique gear. nespp4_FLT and nespp4_FLT_DESC are labels assigned to gear for accounting purposes as in BIOSTAT)}
  \item{oc - Open/closed trip designation for scallop trips, default = "ALL" (other fleets would be considered open which is included in "ALL")}
  \item{trp - Trip access type for scallop trips, options include "GEN", "LIM" and "ALL", default = "ALL"}
  \item{start_year - Start year for calculations, no default. "post_CAMS" can start in 2018 but may start later if discards manually processed during CAMS transition}
  \item{end_year - End year for calculations no default. Generally the end year for "post_CAMS" data is the terminal year of the assessment}
  \item{nespp4_FLT_DESC - A vector of fleet descriptions corresponding to nespp4_FLT}
}}

\item{tab_age_data_source}{A list providing the following information to populate the AGE_DATA_SOURCE tab:
\itemize{
  \item{YEAR - Vector of years for which age data source provided, likely match data_years argument}
  \item{AGE_DATA_SOURCE - A vector of age data source corresponding to YEAR, options include "SVDBS" for survey ages (used for pre-CAMS years), "CFAGEAA" for commercial AA tables, "CAMSAGE" to pull ages directly from CAMS}
}}

\item{outfile}{A string for the final exported file name, does NOT need to include .xlsx extention. No default.

NOTE: tab_blocks$block_opt = checkLENGTHS will NOT work unless connected to the VPN
NOTE: ALK_HOLES, EXCLUSIONS, and LENGTH_IMPUTATIONS tabs are not populated by this function but placeholders are put in the resulting output file so that they can be filled externally as needed.
NOTE: If you receive the following error, then the .xlsx file by the provided name already exists and can't be overwritten (generally happens when the file is open): [ERROR] workbook_close(): Error creating 'outfile.xlsx'. System error = Permission denied Error: Error in libxlsxwriter: 'Error creating output xlsx file. Usually a permissions error.'
NOTE: Code assumes that all gears fish in all areas when setting up the NESPP4_FLT and SVAGE_DEFINITION tabs, if this is not the case then invalid area/gear combinations can be filtered out of the final table and re-saved manually (REGIONS, NESPP4_FLT, SVAGE_DEFINITION tabs would need to be manually adjusted)
NOTE: Code assumes one region for each area fished, if this is not the case then region can be manually updated in the following tabs: BLOCKS, REGIONS, AGE_DATA_SOURCE
NOTE: The following warning may indicate different lengths for nespp4_flt$nespp4_FLT, nespp4_flt$gear_group, nespp4_flt$NEGEAR_group, nespp4_flt$meshsize_abbrev, or nespp4_flt$nespp4_FLT_DESC objects: 
      Warning message: In cbind(SPECIES_ITIS = as.character(species_itis), STOCK_ABBREV = stock_abbrev,  :
      number of rows of result is not a multiple of vector length
NOTE: If you find NAs in the REGION_ID for the BLOCKS tab, then there was a stat area in the stockeff_pre_prod.mv_cf_stock_data_length_o product that was not provided in the tab_regions$stat_areas argument.}
}
\value{
A list containing the following:
\itemize{
  \item{regions_missing - Table of AREAs that appear in oracle data but not in tab_regions$stat_area argument for this function (may be missing blocking)}
}
}
\description{
Fill OB specification template for stockEff
}
