% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ismooth_functions.R
\name{filltest}
\alias{filltest}
\title{filltest}
\usage{
filltest(data, YearStart, termyearstart, termyearend, myseries, myfilltests)
}
\arguments{
\item{data}{tibble with columns Year, Series, Index}

\item{YearStart}{data filtered to years >= this value}

\item{termyearstart}{first year of terminal years to be used in analysis}

\item{termyearend}{last year of terminal years to be used in analysis}

\item{myseries}{vector of series names in quotes that subsets the Series column}

\item{myfilltests}{tibble with columns relYear, Series, Howfill where relYear is the year relative to the terminal year in each evaluation}
}
\value{
tibble with columns TermYear (terminal year for that evaluation), Mult_All (Ismooth multiplier using all data), Treatment (either Filled or Missing), Multiplier (Ismooth multiplier for that Treatment), error (Multiplier - Mult_All)
}
\description{
compare full time series of data with missing and filled time series to see which is preferred
}
