% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOO.R
\name{LOO}
\alias{LOO}
\title{Leave-one-out}
\usage{
LOO(mod = NULL, dosdrep = TRUE)
}
\arguments{
\item{mod}{An rds from a WHAM fit}

\item{dosdrep}{TRUE - whether or not to calculate standard deviations}
}
\value{
A named list with each WHAM fit, where the name is of the index dropped
}
\description{
refits a WHAM model sequentially dropping one index at a time
}
\examples{
mod=readRDS("insert path to a WHAM.rds")
LOO.mods=LOO(mod=mod,dosdrep=TRUE)
You can add the base model to LOO.mods after the fact LOO.mods[["base"]]=mod
compare_wham_models can work to create comparison plots
}
