% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePresentation.R
\name{makePresentation}
\alias{makePresentation}
\title{makePresentation}
\usage{
makePresentation(
  plotStorage = NULL,
  figureExtension = ".png",
  slideTitle = "Supplemental figures",
  format = "pptx",
  outdir = NULL,
  filename = NULL
)
}
\arguments{
\item{plotStorage}{A file path to the folder containing all figures to include in a powerpoint, this folder can contain sub-directories whose images will also be included}

\item{figureExtension}{A string specifying the file extension for the type of figures to include, default = ".png"}

\item{slideTitle}{A title for the PowerPoint slides, default = "Supplemental figures"}

\item{format}{A string indicating the type of format to use ("pptx" or "revealjs"), default = "pptx"}

\item{outdir}{A file path for the directory where the final presentation file should be saved, no default.}

\item{filename}{A string for the final presentation file name, no default.}
}
\description{
Generic R code to pull all figures from a folder into a powerpoint. By default this code generates a PowerPoint presentation including all .png files in the specified plotStorage directory. Slides are labeled numerically
}
