% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mdeaths.R
\name{Mdeaths}
\alias{Mdeaths}
\title{Calculate the amount of fish dying given M}
\usage{
Mdeaths(mod = NULL)
}
\arguments{
\item{mod}{An rds from a WHAM fit}
}
\value{
A dataframe with two columns. Year and Deaths(000s mt)
}
\description{
Calculates the amount (000s metric tons) of fish dying given the assumed or estimated M; uses Jan-1 biomass, so an approximation
}
\examples{
mod.dir="m168"
write.dir <- paste("C:/Herring/2022 Assessment/Assessments/WHAM",mod.dir,sep="/")
setwd(write.dir)
mod <- readRDS(paste0(mod.dir,".rds"))
m168deaths=Mdeaths(mod=mod)
plot(m168deaths$Year,m168deaths$Deaths,type='o',ylab="Dead Fish (000's mt)",xlab="")
}
