% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ismooth_functions.R
\name{plot_6_ts}
\alias{plot_6_ts}
\title{plot_6_ts}
\usage{
plot_6_ts(data, mytitle)
}
\arguments{
\item{data}{tibble with columns Year, Series, Index}

\item{mytitle}{character string that will appear at top of plot}
}
\value{
ggplot object than can be seen using print() command
}
\description{
creates standardized plot of tibble resulting from clean_stockeff_survey_data function
}
