% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ismooth_functions.R
\name{plot_filltest}
\alias{plot_filltest}
\title{plot_filltest}
\usage{
plot_filltest(res, mytitle, printplots = FALSE)
}
\arguments{
\item{res}{tibble resulting from filltest function}

\item{mytitle}{character string to appear at top of plot}

\item{printplots}{Boolean whether to automatically print plots (default=FALSE)}
}
\value{
list with three plots and root mean square error between multipliers using all data and treatments
}
\description{
create three plots from results of filltest function: XY, Difference Distribution, Difference Time Series
}
