% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ismooth_functions.R
\name{plot_standardized_and_avg}
\alias{plot_standardized_and_avg}
\title{plot_standardized_and_avg}
\usage{
plot_standardized_and_avg(data, mytitle)
}
\arguments{
\item{data}{list with std_data and avg tibbles}

\item{mytitle}{character string to appear at top of plot}
}
\value{
ggplot object that can be seen using print() command
}
\description{
plots the results of standardize_and_combine_indices function in a single panel with all series and average labeled
}
