% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNAAfxn.R
\name{plotNAAfxn}
\alias{plotNAAfxn}
\title{Plot NAA devs from WHAM fit}
\usage{
plotNAAfxn(mods = NULL, cor = NULL)
}
\arguments{
\item{mods}{An rds from a WHAM fit with NAA random effects}

\item{cor}{NULL - a placeholder for future labeling}
}
\value{
A ggplot. Must use ggsave after running the function to save the plot
}
\description{
Plot NAA devs from WHAM fit
}
\examples{
naaplot=plotNAAfxn(mods=mod,cor="NA")
ggsave(paste("YOURNAME","_NAAdevs.jpeg"),naaplot,path=YOURPATH)
}
