% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P-star.basic.table.R
\name{pstartable}
\alias{pstartable}
\title{Conduct the P* projections for the MAFMC}
\usage{
pstartable(
  mod = NULL,
  SSBmsy = NULL,
  catch.year1 = NULL,
  projyr = 3,
  CV = 1.5,
  avg.abc = NULL
)
}
\arguments{
\item{mod}{An rds from a WHAM fit}

\item{SSBmsy}{the SSBmsy proxy used to calculate ratio between biomass and reference point}

\item{catch.year1}{the amount of catch in the assessment year (usually the ABC)}

\item{CV}{the level of uncertainty}

\item{avg.abc}{average ABC, if using, specify, otherwise don't specify}

\item{proj.yr}{how many years to project into the future}
}
\value{
A table containing:
\itemize{
  \item{Year - projection Year}
  \item{OFL - overfishing limit}
  \item{ABC - allowable biological catch}
  \item{SSB/SSBMSY - ratio between SSB in that year and reference point}
  \item{F - fishing mortality}
  \item{SSB - spawning stock biomass}
  \item{P* - Pstar value}
}
}
\description{
Conducts the P* projections, which incorporates uncertainty and risk in ABC calculation from OFL using a fitted WHAM model
}
\examples{
mod <- readRDS("Modeling/Run3/fit.RDS")
SSBmsy <- 11225        
catch.year1 <- 7557
CV <- 60 
pstar60 <- pstar(mod,SSBmsy,catch.year1,CV=CV)
write.csv(pstar60, "Projections/Pstar60.csv", row.names = FALSE)
}
\author{
E Liljestrand 7-19-2024, expanded code provided by C Adams for butterfish 7-5-2022
}
