% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullResults_WHAM.R
\name{pullResults_WHAM}
\alias{pullResults_WHAM}
\title{Pull results from WHAM}
\usage{
pullResults_WHAM(model = NULL, multiWHAM = FALSE)
}
\arguments{
\item{model}{A fitted WHAM model object, no default.}

\item{multiWHAM}{A boolean, if TRUE, assumes that the model is generated by multi-WHAM and includes additional indexing for region/stock as well as any differences in labels. Default = FALSE.}
}
\value{
A list containing the following
\itemize{
  \item{brps - A table of BRP estimates with 95\% CI for F, SSB and MSY proxies}
  \item{NAA - A table of NAA}
  \item{ FAA - A table of FAA}
  \item{Mohns_rho - A list of Mohn's rho values for F, SSB, and NAA}
  \item{F.yr - A table containing a time series of F estimates and CVs}
  \item{F.yr_adj - F.yr table with Mohn's rho adjustments}
  \item{SSB.yr - A table containing a time series of SSB estimates and CVs}
  \item{SSB.yr_adj - SSB.yr table with Mohn's rho adjustments}
  \item{Rect.yr - A table containing a time series of Recruitment and CVs}
  \item{Rect.yr_adj - Rect.yr table with Mohn's rho adjustments}
  \item{termyr.ests.cis - A table of terminal year estimates of F, SSB, and Recruitment with 95\% and 90\% CIs}
}
}
\description{
Pull out results required for MT report writing from a WHAM model
}
