% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stockEff.R
\name{read_stockEff}
\alias{read_stockEff}
\title{read_stockEff}
\usage{
read_stockEff(
  doLogin = FALSE,
  species_itis = NULL,
  stock_abbrev = NULL,
  sex_type = "NONE",
  mode = "prod",
  module = NULL,
  product = NULL,
  outdir = here::here(),
  saveout = TRUE,
  outname = NULL
)
}
\arguments{
\item{doLogin}{A boolean, if TRUE then function will ask for login credentials, if FALSE then function establishes connection using keyring credentials and updated httr2 package, Default = FALSE}

\item{species_itis}{6 digit ITIS identifying the species for which data should be pulled, no default.}

\item{stock_abbrev}{A string describing the stock abbreviation assigned by stock efficiency (this shows up in the web address for STOCKEFF products), no default.}

\item{sex_type}{A string describing the sex type for which data should be pulled (this shows up in the web address for STOCKEFF products), default = "NONE". Common options include "NONE", "MALE", "FEMALE", "UNSEXED"}

\item{mode}{A string indicating production "prod" or pre-production "test" modes, default = "prod"}

\item{module}{A string indicating the module for which products will be pulled, no default. Options include:
\itemize{
  \item{"survey" - Correspond to SV tab in stockEff}
  \item{"commercial" - Correspond to CF tab in stockEff}
  \item{"observer" - Correspond to OB tab in stockEff}
  \item{Nothing yet available for MRIP tab (as of 3/6/24)}
}}

\item{product}{A vector of strings corresponding to the CSV files available in the SV modual of STOCKEFF, names should NOT need to include ".csv" extensions. No default.}

\item{outdir}{A directory where temporary files will be stored during processing (deleted when function finishes running), and final RDS object will be saved. Default to here::here().}

\item{saveout}{A boolean, if TRUE also provide outname argument for saved copy of output, otherwise results returned to R but not saved, default = TRUE.}

\item{outname}{A string to use as the output file name (will always be time stamped, do NOT add file extension to name), default = paste0("stockEff_",module) with timestamp.}
}
\value{
An R list object where each item in the list corresponds to a product that was pulled from stockEff and an additional date_pulled object with the time stamp. If saveout = TRUE then this object is also saved as an RData file. When returned RData file is loaded into R it will automatically be named "stockEff_storage"
}
\description{
Function reads stockEff products into R and saves as a time-stamped list.
Options allow data to be pulled for different species, stocks, production status, modules, and sex types.
This is essentially an R function wrapper around the "Connecting to StockEff via R" example user guide in confluence.
}
\examples{
# GB cod survey pre-production survey example
read_stockEff(species_itis = 164712, stock_abbrev = "GBK", sex_type = "NONE", module = "survey", mode = "test", product = c("strat_mean", "strat_mean_length"), outname = "example_GBcod_survey")
# Female spiny dogfish production survey example
read_stockEff(species_itis = 160617, stock_abbrev = "UNIT", sex_type = "FEMALE", module = "survey", mode = "prod", product = c("strat_mean", "strat_mean_length"), outname = "example_FEMALEdogfish_survey")
# Butterfish production CF example
read_stockEff(doLogin = FALSE, species_itis = 172567, stock_abbrev = "UNIT", sex_type = "NONE", module = "commercial", mode = "prod", product = c("landings_gear", "bio_data_summary"), outname = "example_butterfish_CF")

NOTE:
- You must have the httr package loaded to run this function if you are not using the pdbCollaborations package implementation
- The function cannot query assessment_info.csv products (you will get the following error if you try: "Error in read.table ... more columns than column names")
}
