% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_selftestplot.R
\name{selftesterrorfxn}
\alias{selftesterrorfxn}
\title{multi_selftestplot}
\usage{
selftesterrorfxn(metric = NULL, mod = NULL, selftest = NULL)
}
\arguments{
\item{mod}{result of fit_wham.}

\item{selftest}{result of self_test applied to mod.}

\item{pngprefix}{(default=NULL) prefix added to default names for all saved png files.}

\item{save.direct}{directory where you want png plots saved.}
}
\value{
saves 3 png files, one each for F, recruitment, and SSB, to the specified directory.
}
\description{
Automated creation of plots from a (multi)WHAM simulation self-test. Main plot shows mean, median, and 95%CI. 
Relative error = (Est-True)/True computed for F, R, and SSB.
The multi_selftestplot requires another function, selftesterrorfxn, which is included below. Users should not need to explicitly call the selftesterrorfxn. It gets called within multi_selftestplot.
}
\examples{
mod=fit_wham(input,...)
selftest=self_test(mod,...)
multi_selftestplot(mod=mod,selftest = selftest,pngprefix = "pngname",save.direct = "C:/yourdirectory/here")

}
