% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ismooth_functions.R
\name{standardize_and_combine_indices}
\alias{standardize_and_combine_indices}
\title{standardize_and_combine_indices}
\usage{
standardize_and_combine_indices(
  data,
  YearStart,
  YearEnd,
  myseries,
  mynarm = FALSE
)
}
\arguments{
\item{data}{tibble with columns Year, Series, Index}

\item{YearStart}{data filtered to years >= this value}

\item{YearEnd}{data filtered to years <= this value}

\item{myseries}{vector of series names in quotes that subsets the Series column}

\item{mynarm}{Boolean to remove NA values when calculating annual average index, default=FALSE}
}
\value{
list with std_data tibble that adds Index_Standardized column to data and avg tibble the has columns Year and avg
}
\description{
standardizes data by dividing by mean during defined time period and averages the series
}
